Adds sneaky sneak views
Adds some views behind the scrollview, of the same colour as the top and
bottom sections.

This means that pulling the scrollview past its limits no longer shows
white, instead showing a colour matching the top section (when pulling
above the top) and the bottom section (when pulling below the bottom).